perf(matcher): LIMIT seeds lazily — 10,000 eager PatternMatches to answer LIMIT 100 are gone
push_limit_into_match pushing WAS the defect's amplifier: source_cap =
100x the limit seeded 10k PatternMatch (Vec + String each) eagerly —
grow_one + drop_glue were 60% of PatternExecutor::execute on the tracked
cell. Seeds now build as the expansion reaches them under a cap;
uncapped stays eager (consumes them all; the parallel branch zips).
Tracked cell 322 -> 95 us (-70.6%); LIMIT 10 shape -22 -> 13.5 us; the
unrelated two_edge_distinct_filtered_path[ENDS WITH] straddler moved
-22.4% as a bonus; worst tracked mover +7.6%, controls flat. The lazy
path had ZERO Rust coverage (a wrong seed index left 1894 tests green) —
matcher_limit_seed_tests pins matching starts placed behind
non-matching. Hop expansion extracted to matcher_expansion.rs
(matcher.rs 2502->1981; execute's exception REMOVED as stale;
add_connections' hand-tightened).
