+++ crates/kglite/src/graph/core/pattern_matching/matcher.rs
@@ -455,462 +455,6 @@ impl<'a> PatternExecutor<'a> {
         self
     }
 
-    /// Choose and materialise the start-node set for `pattern`, capped at
-    /// `source_cap`. Runs once per [`Self::execute`] call, before the
-    /// expansion loop — never per row.
-    fn seed_start_nodes(
-        &self,
-        pattern: &Pattern,
-        first_node: &NodePattern,
-        has_edges: bool,
-        source_cap: Option<usize>,
-    ) -> Result<Vec<NodeIndex>, String> {
-        // Pre-bound first nodes (e.g. `MATCH (f {id: X}) MATCH (f)-[:R]->(c)`)
-        // must skip the inverted-index fast path — that path returns every source
-        // for the edge type, ignoring the binding. find_matching_nodes resolves
-        // the variable to a single node directly.
-        let first_is_prebound = first_node
-            .variable
-            .as_ref()
-            .map(|v| self.pre_bindings.get(v).is_some())
-            .unwrap_or(false);
-
-        // Try connection-type inverted index for untyped source nodes with typed edges.
-        // Instead of iterating all 124M nodes hoping to find P31 sources, the inverted
-        // index gives us exactly which nodes have P31 outgoing edges.
-        let mut initial_nodes = if !first_is_prebound
-            && has_edges
-            && first_node.node_type.is_none()
-            && first_node.properties.is_none()
-        {
-            // Check if the first edge has connection type(s) we can look up.
-            // `[:A|B]` needs the sources of EVERY branch: taking only the
-            // singular `connection_type` dropped every start node whose sole
-            // matching edge was on a later branch.
-            let edge_conn_types: Option<Vec<InternedKey>> =
-                if let Some(PatternElement::Edge(ep)) = pattern.elements.get(1) {
-                    if ep.var_length.is_none() {
-                        match ep.conn_filter() {
-                            ConnTypeFilter::Any => None,
-                            ConnTypeFilter::One(key) => Some(vec![key]),
-                            ConnTypeFilter::AnyOf(keys) => Some(keys),
-                        }
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                };
-            // Check edge direction — inverted index only covers outgoing sources
-            let is_outgoing = if let Some(PatternElement::Edge(ep)) = pattern.elements.get(1) {
-                ep.direction == EdgeDirection::Outgoing
-            } else {
-                false
-            };
-            if let (Some(conn_types), true) = (edge_conn_types, is_outgoing) {
-                // Pass `source_cap` through so we don't eagerly copy the
-                // whole 400 MB source list from the inverted index for
-                // a query that only needs 1 000 of them. An alternation
-                // unions the per-branch source lists; if the index is
-                // unavailable for ANY branch the union would be short, so
-                // the whole lookup falls back to the full node scan.
-                let mut union: Vec<NodeIndex> = Vec::new();
-                let mut complete = true;
-                for ct in conn_types {
-                    match self
-                        .graph
-                        .graph
-                        .sources_for_conn_type_bounded(ct, source_cap)
-                    {
-                        Some(sources) => {
-                            union.extend(sources.into_iter().map(|s| NodeIndex::new(s as usize)));
-                        }
-                        None => {
-                            complete = false;
-                            break;
-                        }
-                    }
-                }
-                if complete {
-                    // A node can source more than one branch; the caller
-                    // treats each start node once.
-                    union.sort_unstable();
-                    union.dedup();
-                    union
-                } else {
-                    self.find_matching_nodes(first_node)?
-                }
-            } else {
-                self.find_matching_nodes(first_node)?
-            }
-        } else {
-            self.find_matching_nodes(first_node)?
-        };
-        if let Some(cap) = source_cap {
-            if initial_nodes.len() > cap {
-                initial_nodes.truncate(cap);
-            }
-        }
-        Ok(initial_nodes)
-    }
-
-    /// Execute the pattern and return all matches
-    pub fn execute(&self, pattern: &Pattern) -> Result<Vec<PatternMatch>, String> {
-        if pattern.elements.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        // Start with the first node pattern
-        let first_node = match &pattern.elements[0] {
-            PatternElement::Node(np) => np,
-            _ => {
-                return Err(
-                    "Pattern must start with a node in parentheses. Example: (n:Person) or ()"
-                        .to_string(),
-                )
-            }
-        };
-
-        // Find all nodes matching the first pattern.
-        // For multi-hop patterns with max_matches, cap the source candidates to avoid
-        // O(N) allocation when only a small number of results are needed (e.g. LIMIT 10
-        // on an 11M-node type). The expansion loop enforces the exact max_matches.
-        let has_edges = pattern.elements.len() > 1;
-        let source_cap = if has_edges {
-            // Multi-hop with LIMIT: cap sources to avoid O(N) allocation + PatternMatch
-            // construction for millions of nodes. The expansion loop enforces exact
-            // max_matches via early-exit. 100x headroom handles sparse match patterns
-            // (each source needs only a 1% chance of producing a match to hit the limit).
-            self.max_matches.map(|m| m.saturating_mul(100).max(1000))
-        } else {
-            // Single-node pattern: exact truncation
-            self.max_matches
-        };
-        let initial_nodes = self.seed_start_nodes(pattern, first_node, has_edges, source_cap)?;
-
-        // Initialize matches with first node bindings
-        let mut matches: Vec<PatternMatch> = initial_nodes
-            .iter()
-            .map(|&idx| {
-                let mut pm = PatternMatch {
-                    bindings: Vec::new(),
-                    exact_path: None,
-                };
-                if let Some(ref var) = first_node.variable {
-                    pm.bindings.push((var.clone(), self.node_to_binding(idx)));
-                }
-                pm
-            })
-            .collect();
-
-        // Track current node indices for each match
-        let mut current_indices: Vec<NodeIndex> = initial_nodes;
-
-        // Pre-allocate dedup set for distinct_target_var optimization
-        let mut distinct_seen: HashSet<NodeIndex> = if self.distinct_target_var.is_some() {
-            HashSet::with_capacity(matches.len())
-        } else {
-            HashSet::new()
-        };
-
-        // Process edge-node pairs
-        let mut i = 1;
-        while i < pattern.elements.len() {
-            // max_matches is enforced DURING expansion (inner-loop checks below),
-            // not between hops, to avoid breaking before edges are expanded.
-            let is_last_hop = i + 2 >= pattern.elements.len();
-            if let Some(msg) = self.interrupt_reason() {
-                return Err(msg);
-            }
-
-            let edge_pattern = match &pattern.elements[i] {
-                PatternElement::Edge(ep) => ep,
-                _ => return Err("Expected edge pattern after node. Use -[:TYPE]-> for outgoing, <-[:TYPE]- for incoming.".to_string()),
-            };
-
-            i += 1;
-            if i >= pattern.elements.len() {
-                return Err("Edge pattern must be followed by a node pattern. Example: ()-[:KNOWS]->(n:Person)".to_string());
-            }
-
-            let node_pattern = match &pattern.elements[i] {
-                PatternElement::Node(np) => np,
-                _ => return Err("Expected node pattern after edge. Complete the pattern with a node: ()-[:EDGE]->(node)".to_string()),
-            };
-
-            // Internal binding names are invariant for the whole hop. Build
-            // them once rather than formatting a fresh String for every
-            // matched relationship on the expansion hot path.
-            let anonymous_var_path_binding = (edge_pattern.variable.is_none()
-                && edge_pattern.needs_path_info
-                && edge_pattern.var_length.is_some())
-            .then(|| format!("__anon_vlpath_{i}"));
-            let track_fixed_trail =
-                edge_pattern.var_length.is_none() && edge_pattern.needs_path_info;
-
-            // Expand each current match
-            let (mut new_matches, mut new_indices) = if matches.len() >= EXPANSION_RAYON_THRESHOLD
-                && self.max_matches.is_none()
-            {
-                // Parallel expansion — each match's expand_from_node is independent.
-                // Errors (e.g. deadline exceeded) are captured via AtomicBool and
-                // the first error message is saved for propagation after the parallel section.
-                let had_error = std::sync::atomic::AtomicBool::new(false);
-                let first_error: std::sync::Mutex<Option<String>> = std::sync::Mutex::new(None);
-                let results: Vec<(PatternMatch, NodeIndex)> = matches
-                    .par_iter()
-                    .zip(current_indices.par_iter())
-                    .flat_map(|(current_match, &source_idx)| {
-                        // Short-circuit once any thread has detected a timeout/error,
-                        // and independently check the deadline from each thread so a
-                        // parallel expansion over 100M+ sources cannot run unbounded.
-                        if had_error.load(std::sync::atomic::Ordering::Relaxed) {
-                            return Vec::new();
-                        }
-                        if let Some(msg) = self.interrupt_reason() {
-                            if !had_error.swap(true, std::sync::atomic::Ordering::Relaxed) {
-                                *first_error.lock().unwrap() = Some(msg);
-                            }
-                            return Vec::new();
-                        }
-                        let expansions = match self.expand_from_node(
-                            source_idx,
-                            edge_pattern,
-                            node_pattern,
-                            None,
-                            self.bound_target(node_pattern, current_match),
-                        ) {
-                            Ok(exp) => exp,
-                            Err(e) => {
-                                if !had_error.swap(true, std::sync::atomic::Ordering::Relaxed) {
-                                    *first_error.lock().unwrap() = Some(e);
-                                }
-                                return Vec::new();
-                            }
-                        };
-                        expansions
-                            .into_iter()
-                            .filter_map(|(target_idx, edge_binding)| {
-                                if reuses_bound_relationship(current_match, &edge_binding) {
-                                    return None;
-                                }
-                                if let Some(ref var) = node_pattern.variable {
-                                    if let Some(&bound_idx) = self.pre_bindings.get(var) {
-                                        if target_idx != bound_idx {
-                                            return None;
-                                        }
-                                    }
-                                    // Enforce intra-pattern variable constraint
-                                    let already_bound = current_match.bindings.iter().find_map(
-                                        |(name, binding)| {
-                                            if name == var {
-                                                match binding {
-                                                    MatchBinding::Node { index, .. }
-                                                    | MatchBinding::NodeRef(index) => Some(*index),
-                                                    _ => None,
-                                                }
-                                            } else {
-                                                None
-                                            }
-                                        },
-                                    );
-                                    if let Some(bound_idx) = already_bound {
-                                        if target_idx != bound_idx {
-                                            return None;
-                                        }
-                                    }
-                                }
-                                let mut new_match = current_match.clone();
-                                if track_fixed_trail {
-                                    extend_fixed_trail(&mut new_match, &edge_binding);
-                                }
-                                if let Some(ref var) = edge_pattern.variable {
-                                    new_match.bindings.push((var.clone(), edge_binding));
-                                } else if let Some(ref internal_var) = anonymous_var_path_binding {
-                                    new_match
-                                        .bindings
-                                        .push((internal_var.clone(), edge_binding));
-                                }
-                                if let Some(ref var) = node_pattern.variable {
-                                    new_match
-                                        .bindings
-                                        .push((var.clone(), self.node_to_binding(target_idx)));
-                                }
-                                Some((new_match, target_idx))
-                            })
-                            .collect::<Vec<_>>()
-                    })
-                    .collect();
-                // Propagate any error that occurred during parallel expansion
-                if had_error.load(std::sync::atomic::Ordering::Relaxed) {
-                    let err = first_error
-                        .into_inner()
-                        .unwrap()
-                        .unwrap_or_else(|| "parallel expansion failed".to_string());
-                    return Err(err);
-                }
-                // Apply distinct-target dedup for parallel results (sequential path
-                // does this inline, but parallel path can't without synchronization).
-                let needs_dedup = i + 2 >= pattern.elements.len()
-                    && self
-                        .distinct_target_var
-                        .as_ref()
-                        .is_some_and(|dtv| node_pattern.variable.as_deref() == Some(dtv.as_str()));
-                if needs_dedup {
-                    let mut seen_targets = HashSet::new();
-                    let filtered: Vec<_> = results
-                        .into_iter()
-                        .filter(|(_, target_idx)| seen_targets.insert(*target_idx))
-                        .collect();
-                    filtered.into_iter().unzip()
-                } else {
-                    results.into_iter().unzip()
-                }
-            } else {
-                // Sequential expansion with max_matches early-exit
-                let mut new_matches_seq = Vec::new();
-                let mut new_indices_seq = Vec::new();
-                let mut expand_count: usize = 0;
-                // At the last hop, enforce exact max_matches.
-                // At intermediate hops, use a generous overcommit (50x) to avoid
-                // expanding far more intermediates than needed while ensuring
-                // enough survive to produce max_matches final results.
-                let hop_limit = if is_last_hop {
-                    self.max_matches
-                } else {
-                    self.max_matches.map(|m| m.saturating_mul(50).max(1000))
-                };
-                for (current_match, &source_idx) in matches.iter().zip(current_indices.iter()) {
-                    if hop_limit.is_some_and(|max| new_matches_seq.len() >= max) {
-                        break;
-                    }
-                    let remaining = hop_limit.map(|max| max.saturating_sub(new_matches_seq.len()));
-                    let hint = self.bound_target(node_pattern, current_match);
-                    let expansions = self.expand_from_node(
-                        source_idx,
-                        edge_pattern,
-                        node_pattern,
-                        remaining,
-                        hint,
-                    )?;
-                    for (target_idx, edge_binding) in expansions {
-                        if reuses_bound_relationship(current_match, &edge_binding) {
-                            continue;
-                        }
-                        expand_count += 1;
-                        if expand_count.is_multiple_of(1024) {
-                            if let Some(msg) = self.interrupt_reason() {
-                                return Err(msg);
-                            }
-                        }
-                        if hop_limit.is_some_and(|max| new_matches_seq.len() >= max) {
-                            break;
-                        }
-                        if let Some(ref var) = node_pattern.variable {
-                            if let Some(&bound_idx) = self.pre_bindings.get(var) {
-                                if target_idx != bound_idx {
-                                    continue;
-                                }
-                            }
-                            // Enforce intra-pattern variable constraint:
-                            // if this variable was already bound earlier in the
-                            // same pattern, the target must match that binding.
-                            let already_bound =
-                                current_match.bindings.iter().find_map(|(name, binding)| {
-                                    if name == var {
-                                        match binding {
-                                            MatchBinding::Node { index, .. }
-                                            | MatchBinding::NodeRef(index) => Some(*index),
-                                            _ => None,
-                                        }
-                                    } else {
-                                        None
-                                    }
-                                });
-                            if let Some(bound_idx) = already_bound {
-                                if target_idx != bound_idx {
-                                    continue;
-                                }
-                            }
-                        }
-                        // Distinct-target dedup: at the last hop, skip targets already seen
-                        if i + 1 >= pattern.elements.len() {
-                            if let Some(ref dtv) = self.distinct_target_var {
-                                if node_pattern.variable.as_deref() == Some(dtv.as_str())
-                                    && !distinct_seen.insert(target_idx)
-                                {
-                                    continue;
-                                }
-                            }
-                        }
-                        let mut new_match = current_match.clone();
-                        if track_fixed_trail {
-                            extend_fixed_trail(&mut new_match, &edge_binding);
-                        }
-                        if let Some(ref var) = edge_pattern.variable {
-                            new_match.bindings.push((var.clone(), edge_binding));
-                        } else if let Some(ref internal_var) = anonymous_var_path_binding {
-                            new_match
-                                .bindings
-                                .push((internal_var.clone(), edge_binding));
-                        }
-                        if let Some(ref var) = node_pattern.variable {
-                            new_match
-                                .bindings
-                                .push((var.clone(), self.node_to_binding(target_idx)));
-                        }
-                        new_matches_seq.push(new_match);
-                        new_indices_seq.push(target_idx);
-                    }
-                }
-                (new_matches_seq, new_indices_seq)
-            };
-
-            // Check deadline / cancellation after expansion (covers both
-            // parallel and sequential paths)
-            if let Some(msg) = self.interrupt_reason() {
-                return Err(msg);
-            }
-
-            // Apply hop limit truncation (for parallel path which can't early-exit)
-            let truncate_limit = if is_last_hop {
-                self.max_matches
-            } else {
-                self.max_matches.map(|m| m.saturating_mul(50).max(1000))
-            };
-            if let Some(max) = truncate_limit {
-                new_matches.truncate(max);
-                new_indices.truncate(max);
-            }
-
-            // Intermediate dedup: when distinct_target_var is set and this is
-            // NOT the final hop and the current node is anonymous (no variable),
-            // deduplicate by NodeIndex to reduce work at subsequent hops.
-            if self.distinct_target_var.is_some()
-                && i + 1 < pattern.elements.len()
-                && node_pattern.variable.is_none()
-            {
-                let mut seen_idx = HashSet::with_capacity(new_indices.len());
-                let mut deduped_matches = Vec::with_capacity(new_indices.len());
-                let mut deduped_indices = Vec::with_capacity(new_indices.len());
-                for (m, idx) in new_matches.into_iter().zip(new_indices) {
-                    if seen_idx.insert(idx) {
-                        deduped_matches.push(m);
-                        deduped_indices.push(idx);
-                    }
-                }
-                matches = deduped_matches;
-                current_indices = deduped_indices;
-            } else {
-                matches = new_matches;
-                current_indices = new_indices;
-            }
-            i += 1;
-        }
-
-        Ok(matches)
-    }
-
     /// Public wrapper for find_matching_nodes (used by Cypher executor for shortestPath)
     pub fn find_matching_nodes_pub(&self, pattern: &NodePattern) -> Result<Vec<NodeIndex>, String> {
         self.find_matching_nodes(pattern)
@@ -2423,8 +1967,15 @@ impl<'a> PatternExecutor<'a> {
     }
 }
 
+#[path = "matcher_expansion.rs"]
+mod expansion;
+
 #[cfg(test)]
 #[path = "matcher_id_lookup_tests.rs"]
 mod id_lookup_tests;
+
+#[cfg(test)]
+#[path = "matcher_limit_seed_tests.rs"]
+mod limit_seed_tests;
 
 // ============================================================================
